renamed /types/index.js -> /types/types.js
