+++ src/components/app/app.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import MainPage from "../main-page/main-page";
-import {appType} from "../types/index";
+import {appType} from "../types/types";
 
 const App = (props) => {
   const {name, genre, releaseDate, genres, movies} = props;
