+++ src/components/main-page/modules/genres-list-item.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import {listItemType} from "../../types/index";
+import {listItemType} from "../../types/types";
 
 const GenresListItem = (props) => {
   const {genre} = props;
